requirements.txt
asgiref==3.6.0
Django==4.1.9
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
Markdown==3.4.3
mongoengine==0.27.0
pymongo==3.12.3
pytz==2023.3
sqlparse==0.2.4
tzdata==2023.3